cmx4mlops/cmx4mlops/repo/requirements.txt
cmind>=2.0.1
pyyaml
requests
setuptools
giturlparse
tabulate
